Move CLI cli_output module to its own crate
