+++ cli/src/cluster_query.rs
@@ -1,15 +1,17 @@
 use crate::{
     cli::{CliCommand, CliCommandInfo, CliConfig, CliError, ProcessResult},
-    cli_output::*,
-    display::{
-        format_labeled_address, new_spinner_progress_bar, println_name_value, println_transaction,
-    },
     spend_utils::{resolve_spend_tx_and_check_account_balance, SpendAmount},
 };
 use clap::{value_t, value_t_or_exit, App, AppSettings, Arg, ArgMatches, SubCommand};
 use console::{style, Emoji};
 use solana_clap_utils::{
     commitment::commitment_arg, input_parsers::*, input_validators::*, keypair::DefaultSigner,
+};
+use solana_cli_output::{
+    display::{
+        format_labeled_address, new_spinner_progress_bar, println_name_value, println_transaction,
+    },
+    *,
 };
 use solana_client::{
     pubsub_client::PubsubClient,
