+++ cli/src/lib.rs
@@ -18,14 +18,11 @@ macro_rules! pubkey {
     };
 }
 
-#[macro_use]
 extern crate serde_derive;
 
 pub mod checks;
 pub mod cli;
-pub mod cli_output;
 pub mod cluster_query;
-pub mod display;
 pub mod nonce;
 pub mod offline;
 pub mod spend_utils;
