+++ cli/src/offline/mod.rs
@@ -1,6 +1,6 @@
-use crate::cli_output::{CliSignOnlyData, OutputFormat};
 use serde_json::Value;
 use solana_clap_utils::keypair::presigner_from_pubkey_sigs;
+use solana_cli_output::{CliSignOnlyData, OutputFormat};
 use solana_sdk::{
     hash::Hash,
     pubkey::Pubkey,
