+++ cli/src/stake.rs
@@ -4,7 +4,6 @@ use crate::{
         log_instruction_custom_error, CliCommand, CliCommandInfo, CliConfig, CliError,
         ProcessResult,
     },
-    cli_output::{CliStakeHistory, CliStakeHistoryEntry, CliStakeState, CliStakeType},
     nonce::check_nonce_account,
     offline::return_signers,
     spend_utils::{resolve_spend_tx_and_check_account_balances, SpendAmount},
@@ -19,6 +18,7 @@ use solana_clap_utils::{
     offline::*,
     ArgConstant,
 };
+use solana_cli_output::{CliStakeHistory, CliStakeHistoryEntry, CliStakeState, CliStakeType};
 use solana_client::{
     blockhash_query::BlockhashQuery, nonce_utils, rpc_client::RpcClient,
     rpc_request::DELINQUENT_VALIDATOR_SLOT_DISTANCE,
