+++ ledger-tool/src/main.rs
@@ -118,7 +118,7 @@ fn output_entry(
                     })
                     .map(|transaction_status| transaction_status.into());
 
-                solana_cli::display::println_transaction(
+                solana_cli_output::display::println_transaction(
                     &transaction,
                     &transaction_status,
                     "      ",
